refactor exec encode
